core: fix directory scanning by returning actual entry type
Directory scanning may broke for users that filter entries based
on entry type, like list only regular files.

The root of all evil is that we were retrieving entry type as if
getdents() were used, but getdents64() has a type field defined
in its own entry structure which must be used for this purpose
*and* sizeof() incorrectly used on work's array field because of
indirection.

We were returning mostly garbage for type which would usually be
unknown so it could luckily work (which is the reason it took
us so long to find it) but it could happen that a different type
would be returned which would lead to incorrect behavior when
scanning the directory.

ScyllaDB, for example, failed to boot because an essential file
was skipped due to filtering based on type (look:
https://github.com/scylladb/scylla/issues/3274).

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>
Message-Id: <20180312183901.4067-1-raphaelsc@scylladb.com>