core::reactor: move task_quota_timer_thread shutdown to reactor::stop()
Make sure that the task_quota_timer_thread is always explicitly shutdown.
Shutting it down from the reactor's destructor doesn't ensure that because
reactor is not always explicitly destroyed.
When it doesn't - since the reactor's instance is allocated on a heap and
then its pointer is stored in the reactor_holder which is effectively a
global unique_ptr the reactor, it will be automatically destroyed in a
hard-to-predict order compared to other global variables which are accessed
from the task_quota_timer_thread, like g_need_preempt.

Fixes #3345

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>