perftune.py: allow explicitly giving a CPU mask to be used for binding IRQs
In addition to original configuration modes allow explicitly providing
a CPU mask that will be used to distribute IRQs of configured devices
among them.

This is going to be done using a new parameter: --irq-cpu-set (irq_cpu_set in YAML).

This parameter may not be given together with --mode parameter due to
obvious reasons.

Fixes #504

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>