perftune.py: introduce a --get-cpu-mask-quiet parameter
When this new parameter is given perftune.py is going to print the
compute CPU set similarly to --get-cpu-mask if tuning is possible.

If tuning is impossible (the compute mask turns out to be all-zeros) then
if --cpu-mask is not given the full CPU set of the system is printed or
the value given with --cpu-mask is printed otherwise.
Exit status is going to be zero.

--get-cpu-mask would still print an error and return a non-zero exit
status in the later case.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>