net: split Rx checksum features
Separate Rx checksum features into IPv4, TCP and UDP checksum features.

Up until now we were using a shortcut where we always demanded all 3
to be enabled or disabled together.

We saw that with virtio it's not always the case, so it's time we
split them and handle them independently.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>