perftune.py: fix parsing of 'write_back_cache' YAML option
Apparently Python3-yaml prints a boolean values as '1' or '0'
however when it parses those values from the .yaml file it
interprets them as an integer.

Let's fix this by printing whatever value we get from the .yaml
into the string and then convert that string into the 'bool'
using distutils.util.strtobool().

Fixes: #894
Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>