perftune.py: fix a bug in mlx4 IRQs names matching pattern
One should not assume that match groups are going to have the
same index in the pattern like this "some_pattern1(gr1)|some_pattern2(gr2)".

In case of a left side match gr1 will be accessible as match.group(1).
But in case of a right side match gr2 won't be accessible as match.group(1)
but rather as match.group(2).

To avoid this ambiguity we can split the pattern into two by having left and right sides
as separate patterns and apply them separately.

Fixes #895

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>