perftune.py: add VIRTIO fast path IRQs ordering functor
The VIRTIO unlike most other drivers allocates 2 IRQs per queue: one for Rx and one for Tx.
It makes sense to bind both IRQs of the same queue to the same CPU.

IRQs names have the following structure:

virtioXYZ-[input|output].NMK

where XYZ is a device index and NMK - a queue index.

A new functor is going to return the NMK for both input and output IRQs ensuring
these are next to each other in the sorted array of IRQs.

We also pre-sort all IRQs numbers array so that the resulting list of VIRTIO IRQ
is numerically sorted too (since Python's sorting is stable).

As a result since we distribute IRQs evenly among available "IRQ CPUs" and due to a fact that there will always going to be either a single such CPU, or an even number of them, or (in MQ mode) the same numebr of them as the number of VIRTIO queues, we will always bind both Rx and Tx IRQ of the same VIRTIO queue to the same CPU.

Fixes #1165