perftune.py: stop printing 'mode' in --dump-options-file
We want to stop using 'mode' and switch to 'irq_cpu_mask' instead as a more generic
notation.

To achieve that we want to continue support receiving 'mode'/'--mode' as an input
for backward compatibility but stop printing it when '--dump-options-file' is requested
(which is mainly used in order to generate a YAML configuration file to be used later).

Instead, when '--mode' and '--dump-options-file' are given we are going to print a
corresponding 'irq_cpu_mask'.

Fixes #1170