perftune.py: DiskPerfTuner: use self.irqs_cpu_mask for tuning non-NVME disks
Cleanup a forgotten place where we still assume that 'mode' is always defined
or auto-detected.

After introduction of 'irq_cpu_mask' parameter this is no longer the case: we
should always use self.irqs_cpu_mask which is either explicitly defined by a user
or deducted from the 'mode'.

And 'mode' may only be auto-detected if 'irq_cpu_mask' is NOT provided.