perftune.py: auto-select the same number of IRQ cores on each NUMA
Change the current conservative auto-selection logic that allocates
IRQ cores globally and then distributes them evenly (as much as possible)
among NUMA nodes.

This can result in an odd number of IRQ cores, for instance on a 2 NUMA
node system with 48 cores (96 HTs). And as a result NUMA nodes will be unbalanced
since one of them is going to have more compute CPUs, which in line is going
to make them less per-CPU RAM.

This patch makes sure that all NUMA nodes are going to get the same amount of
IRQ cores allocated which in line is going keep compute CPUs balanced too.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>