scripts/perftune.py: make output of --dump-options-file consistent
perftune.py supports a few 'list' configurables and we want such values
to appear in the same order every time we invoke 'perftune.py --dump-options-file ...'
with the same set of arguments.

This is particularly important for tools that detect configuration drift - with this
patch a simple 'diff' can be used.

Fixes #1842